+++ src/emit/types.ts
@@ -474,6 +474,8 @@ export function mapTsType(node: ts.Node, t: ts.Type, checker: ts.TypeChecker): C
         if (sym?.getName() === "TextDecoder") return T_TEXT_DECODER;
         if (sym?.getName() === "FSStats") return T_FS_STATS;
         if (sym?.getName() === "FSDirent") return T_FS_DIRENT;
+        if (sym?.getName() === "FSDirHandleEntry") return T_VALUE;
+        if (sym?.getName() === "FSDir") return T_VALUE;
         if (sym?.getName() === "FSFileHandle") return T_VALUE;
         if (sym?.getName() === "DispatchQueue") return T_DISPATCH_QUEUE;
         if (sym?.getName() === "TemplateStringsArray") return arrayType(T_STRING);